python-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
supabase==2.3.0
langchain-community>=0.0.10
langchain>=0.1.0
langchain-openai==0.0.2
langsmith>=0.0.77
python-jose==3.3.0
requests==2.31.0
aiohttp==3.9.1
openai==1.8.0
pinecone-client==3.0.2 